// schemas.d.ts
export const enum GQLJoinType {
  UNJOIN = 'UNJOIN',
  REQUESTED = 'REQUESTED',
  INVITED = 'INVITED',
  JOINED = 'JOINED',
  HIDDEN = 'HIDDEN'
}
// ...
export const enum GQLNotificationType {
  FRIEND_REQUESTED = 'FRIEND_REQUESTED',
  FRIEND_ACCEPTED = 'FRIEND_ACCEPTED',
  PLAN_JOIN_REQUEST = 'PLAN_JOIN_REQUEST',
  PLAN_JOIN_REQUEST_APPROVED = 'PLAN_JOIN_REQUEST_APPROVED',
  PLAN_INVITE = 'PLAN_INVITE',
  PLAN_INVITE_ACCEPTED = 'PLAN_INVITE_ACCEPTED',
  PLAN_DELETED = 'PLAN_DELETED',
  NEARBY_FRIEND = 'NEARBY_FRIEND',
  PLAN_COMMENT = 'PLAN_COMMENT',
  TAG = 'TAG',
  FUTURE_PLAN_OVERLAP = 'FUTURE_PLAN_OVERLAP',
  LIKED_PLAN = 'LIKED_PLAN',
  LIKED_COMMENT = 'LIKED_COMMENT'
}
// ...
export interface GQLJoin {
  _id?: string;
  _from?: string;
  _to?: string;
  _key?: string;
  joinStatus?: GQLJoinType;
  start?: string;
  end?: string;
}
// ...
export interface MutationToJoinArgs {
  user: string;
  plan: string;
  joinType: GQLJoinType;
  requester?: string;
  start?: string;
  end?: string;
}